+++ src/components/index.js
@@ -1,0 +1,8 @@
+export { Content } from 'components/Content';
+export { Filters } from 'components/Filters';
+export { Footer } from 'components/Footer';
+export { Header } from 'components/Header';
+export { Input } from 'components/Input';
+export { List } from 'components/List';
+export { Main } from 'components/Main';
+export { Todo } from 'components/Todo';
